moved initialising reprod settings to reprod_utils and single quotes in helene.yaml
